% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Check and add any missing columns expected by ICER computation functions}
\usage{
add_missing_columns(.df_check_, .string_columns_, .numerics_columns_)
}
\arguments{
\item{.df_check_}{Dataframe for which the function will ensure existence, and
if required creation, of columns with given names in \code{.string_columns_} and
\code{.numerics_columns_} arguments.}

\item{.string_columns_}{Character vector defining the names of columns of class
\code{character} which the function would check if exist in the \code{.df_check_},
creating the ones that do not exist before binding them to \code{.df_check_}.}

\item{.numerics_columns_}{Character vector  defining the names of columns of
class \code{numeric} which the function would check if exist in the \code{.df_check_},
creating the ones that do not exist before binding them to \code{.df_check_}.}
}
\value{
A dataframe including all the data in \code{.df_check_} in addition to empty
columns with names passed to \code{.string_columns_} and \code{.numerics_columns_}.
}
\description{
This is a utility function that checks and necessary adds
columns to a dataframe.
}
\examples{
\dontrun{
df_test <- data.frame(
  "a" = 1:5,
  "b" = LETTERS[1:5],
  "c" = 6:10
)
df2 <- add_missing_columns(
  .df_check_ = df_test,
  .string_columns_ = c("icer label", "dominance"),
  .numerics_columns_ = c("qalys", "costs")
)
}
}
