% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{backTransform}
\alias{backTransform}
\title{Back transform parameters values}
\usage{
backTransform(.t_data_, .l_params_)
}
\arguments{
\item{.t_data_}{A dataset with data on the transformed scale}

\item{.l_params_}{A list with required parameters' information,
including; names and functions to back transform the parameters to their
original/desired scale.}
}
\value{

}
\description{
Back transform parameters values
}
\examples{
\dontrun{
}
}
