% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca.R
\name{bca}
\alias{bca}
\title{Generates a standard curve for BCA analysis and calculates the concentrations of unknown samples}
\usage{
bca(prot.conc, blank.abs, std.abs, abs, sample.ids, dil.fac)
}
\arguments{
\item{prot.conc}{The concentrations of the protein standards used in the assay.}

\item{blank.abs}{The absorbance value of the blank (buffer only).}

\item{std.abs}{Absorbance values of the standards at the concentrations defined in \code{prot.conc}.}

\item{abs}{Absorbance values of the unknown analyte(s).}

\item{dil.fac}{The dilution factors corresponding to each entered analyte aborbance value.}
}
\description{
Generates a standard curve for BCA analysis and calculates the concentrations of unknown samples
}
\examples{
bca(prot.conc = c(), blank.abs, std.abs = c(), abs = NULL)

}
