% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{calculate_icers}
\alias{calculate_icers}
\title{Calculate the Incremental Cost Effectiveness Ratio ICER(s) and effects and
costs differentials}
\usage{
calculate_icers(
  .df_outcomes_,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
\arguments{
\item{.df_outcomes_}{Dataframe or tibble containing average costs and effects
data.}

\item{.label_effects_}{Character indicating the name of the column containing
the mean values of the effects. Default is \code{"QALYs"}.}

\item{.label_costs_}{Character indicating the name of the column containing
the mean values of the costs. Default is \code{"Costs"}.}
}
\value{
A dataframe or table containing intervention names, mean outcomes
(costs and effects), differential outcomes corresponding to each intervention
and the Incremental Cost Effectiveness Ratios (ICER)s for qualified
comparators.
}
\description{
This function estimates the Incremental Cost Effectiveness Ratio ICER(s).
}
\examples{
\dontrun{
df_outcomes <- data.frame(
  'Intervention' = ShinyPSA::Hyperphosphatemia_PSA$treat \%>\% unname(),
  'QALYs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$e) \%>\% unname(),
  'Costs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$c) \%>\% unname()
)
df_icer <- calculate_icers(
  .df_outcomes_ = df_outcomes,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
}
