% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{calculate_likelihood}
\alias{calculate_likelihood}
\title{Calculate likelihood}
\usage{
calculate_likelihood(
  .samples,
  .func = .func_,
  .args = .args_,
  .l_targets = .l_targets_
)
}
\arguments{
\item{.samples}{A table or vector of sampled parameter values}

\item{.func}{A function defining the model to be calibrated}

\item{.args}{A list of arguments to be passed to .func}

\item{.l_targets}{A list containing a vector of targets' names, a vector
of targets' weights, a vector of targets' distributions, and a table for
each target that contains the values (column name 'value') and standard
errors (column name 'sd') of the corresponding target.}
}
\value{

}
\description{
Calculate likelihood
}
\examples{
\dontrun{
}
}
