% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{calculate_prior}
\alias{calculate_prior}
\title{Calculate prior}
\usage{
calculate_prior(.samples, .l_params = .l_params_)
}
\arguments{
\item{.samples}{A vector/dataset containing sampled/proposed values.}

\item{.l_params}{A list that contains a vector of parameter names,
distributions and distributions' arguments.}
}
\value{

}
\description{
Calculate prior
}
