% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{calculate_priors}
\alias{calculate_priors}
\title{Calculate priors}
\usage{
calculate_priors(.samples, .l_params = .l_params_, .transform = .transform_)
}
\arguments{
\item{.samples}{A vector/dataset containing sampled/proposed values.}

\item{.l_params}{A list that contains a vector of parameter names,
distributions and distributions' arguments.}

\item{.transform}{Logical for whether to back-transform parameters to
their original scale.}
}
\value{

}
\description{
Calculate priors
}
