% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibR_R6.R
\name{calibR_R6}
\alias{calibR_R6}
\title{R6 Class representing a calibR_R6 calibration machine.}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is expected to calibrate a user-defined model
and produce summary plots/tables.
}
\examples{
\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$new`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$sampleR`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$calibrateR_random`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$calibrateR_directed`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$calibrateR_bayesian`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$sample_PSA_values`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$run_PSA`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$summarise_PSA`
## ------------------------------------------------

\dontrun{
}

## ------------------------------------------------
## Method `calibR_R6$draw_plots`
## ------------------------------------------------

\dontrun{
}
Plot Goodness of fit function(s)


## ------------------------------------------------
## Method `calibR_R6$draw_GOF_measure`
## ------------------------------------------------

\dontrun{
}
Plot Targets


## ------------------------------------------------
## Method `calibR_R6$draw_targets_plots`
## ------------------------------------------------

\dontrun{
}
Plot prior and posterior distributions


## ------------------------------------------------
## Method `calibR_R6$draw_distributions_plots`
## ------------------------------------------------

\dontrun{
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calibration_model}}{a Decision-Analytic model under calibration}

\item{\code{calibration_model_args}}{arguments passed to the calibration
model}

\item{\code{calibration_parameters}}{calibration parameters' data}

\item{\code{calibration_targets}}{calibration targets' data}

\item{\code{calibration_results}}{calibration interim results}

\item{\code{transform_parameters}}{boolean for whether to back transform
parameters}

\item{\code{GOF_measure_plot}}{log likelihood values for plot}

\item{\code{model_predictions}}{simulated outputs}

\item{\code{prior_samples}}{samples from parameters' priors}

\item{\code{maximum_a_posteriori}}{parameter set with maximum posterior
probability}

\item{\code{PSA_samples}}{calibration and un-calibration parameters
samples}

\item{\code{PSA_results}}{PSA results}

\item{\code{PSA_summary}}{PSA results' summary}

\item{\code{simulated_targets}}{PSA results}

\item{\code{plots}}{summary plots}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-calibR_R6-new}{\code{calibR_R6$new()}}
\item \href{#method-calibR_R6-sampleR}{\code{calibR_R6$sampleR()}}
\item \href{#method-calibR_R6-calibrateR_random}{\code{calibR_R6$calibrateR_random()}}
\item \href{#method-calibR_R6-calibrateR_directed}{\code{calibR_R6$calibrateR_directed()}}
\item \href{#method-calibR_R6-calibrateR_bayesian}{\code{calibR_R6$calibrateR_bayesian()}}
\item \href{#method-calibR_R6-sample_PSA_values}{\code{calibR_R6$sample_PSA_values()}}
\item \href{#method-calibR_R6-run_PSA}{\code{calibR_R6$run_PSA()}}
\item \href{#method-calibR_R6-summarise_PSA}{\code{calibR_R6$summarise_PSA()}}
\item \href{#method-calibR_R6-draw_plots}{\code{calibR_R6$draw_plots()}}
\item \href{#method-calibR_R6-draw_GOF_measure}{\code{calibR_R6$draw_GOF_measure()}}
\item \href{#method-calibR_R6-draw_targets_plots}{\code{calibR_R6$draw_targets_plots()}}
\item \href{#method-calibR_R6-draw_distributions_plots}{\code{calibR_R6$draw_distributions_plots()}}
\item \href{#method-calibR_R6-clone}{\code{calibR_R6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-new"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-new}{}}}
\subsection{Method \code{new()}}{
Sample prior distribution(s) using one or more sampling method.
This function currently supports: LHS, FGS and RGS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$new(.model, .args, .params, .targets, .transform)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.model}}{The decision analytic model under calibration}

\item{\code{.args}}{Calibration model arguments}

\item{\code{.params}}{Calibration parameters}

\item{\code{.targets}}{Calibration targets}

\item{\code{.transform}}{Logical for whether the model will use
transformed parameters. This allows some functions to back
transform the parameters to their original scale before using them.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Object of class \code{CalibrateR_R6}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-sampleR"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-sampleR}{}}}
\subsection{Method \code{sampleR()}}{
Sample prior distribution(s) using one or more sampling method.
This function currently supports: LHS, FGS and RGS.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$sampleR(.n_samples = 1, .sampling_method, .ssed_no = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.n_samples}}{An integer specifying the number of samples to be
generated.}

\item{\code{.sampling_method}}{The sampling method(s) to use. The function
currently supports LHS, RGS, and FGS.}

\item{\code{.ssed_no}}{random seed number}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Executes the required sampling method and populates the
"prior samples" internal object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-calibrateR_random"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-calibrateR_random}{}}}
\subsection{Method \code{calibrateR_random()}}{
Calibrate the model using one or more random search method(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$calibrateR_random(
  .optim = FALSE,
  .maximise = TRUE,
  .weighted = NULL,
  .sample_method = "LHS",
  .calibration_method = "LLK",
  .calibration_function = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.optim}}{Logical for whether the function is used by an
optimisation algorithm. Default is \code{FALSE}.}

\item{\code{.maximise}}{Logical for whether the output of the function is
used in a maximising optimisation function. Default is \code{TRUE}.}

\item{\code{.weighted}}{Logical for whether the SSE was to be weighted,
default is \code{TRUE}. The weight used by function is
\code{1/(sd^2)}.}

\item{\code{.sample_method}}{The method used to sample from the prior
distribution.}

\item{\code{.calibration_method}}{goodness-of-fit method name.}

\item{\code{.calibration_function}}{goodness-of-fit function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Executes the required calibration method and populates
the samples internal object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-calibrateR_directed"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-calibrateR_directed}{}}}
\subsection{Method \code{calibrateR_directed()}}{
Calibrate the model using one or more directed search method(s).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$calibrateR_directed(
  .gof = "LLK",
  .gof_func = NULL,
  .n_samples = 1,
  .max_iterations = 1000,
  temp = 10,
  trace = NULL,
  .calibration_method = "NM",
  .sample_method = "LHS",
  .maximise = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.gof}}{Name of goodness-of-fit function, default is
log-likelihood.}

\item{\code{.gof_func}}{Goodness-of-fit function; if NULL (default) the
supported function defined by \code{.gof} will be used}

\item{\code{.n_samples}}{Number of Starting sets (gausses) to use.}

\item{\code{.max_iterations}}{Maximum number of algorithm iterations.}

\item{\code{temp}}{SANN algorithm tuning parameter.}

\item{\code{trace}}{Non-negative integer. If positive, tracing information on
the progress of the optimization is produced. Higher values may produce
more tracing information.}

\item{\code{.calibration_method}}{The calibration process.}

\item{\code{.sample_method}}{The method used to sample from the prior
distribution.}

\item{\code{.maximise}}{Boolean for whether the function is to maximise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Executes the required calibration method and populates
the samples internal object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-calibrateR_bayesian"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-calibrateR_bayesian}{}}}
\subsection{Method \code{calibrateR_bayesian()}}{
Calibrate the model using one or more calibration method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$calibrateR_bayesian(
  .b_methods = "SIR",
  .n_resample = 1000,
  .IMIS_iterations = 10,
  .IMIS_sample = 100,
  .MCMC_burnIn = 10000,
  .MCMC_samples = 50000,
  .MCMC_thin = 5,
  .MCMC_rerun = TRUE,
  .diag_ = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.b_methods}}{Bayesian calibration method(s)}

\item{\code{.n_resample}}{Desired number of draws from the posterior}

\item{\code{.IMIS_iterations}}{Maximum number of IMIS iterations}

\item{\code{.IMIS_sample}}{IMIS sample size at each iteration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Executes the required calibration method and populates
the samples internal object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-sample_PSA_values"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-sample_PSA_values}{}}}
\subsection{Method \code{sample_PSA_values()}}{
Sample PSA values for the calibration parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$sample_PSA_values(.calibration_methods, .PSA_samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.calibration_methods}}{Bayesian calibration method(s)}

\item{\code{.PSA_samples}}{Number of PSA sets to sample}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-run_PSA"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-run_PSA}{}}}
\subsection{Method \code{run_PSA()}}{
Run PSA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$run_PSA(.PSA_unCalib_values_ = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.PSA_unCalib_values_}}{PSA values for un-calibrated parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-summarise_PSA"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-summarise_PSA}{}}}
\subsection{Method \code{summarise_PSA()}}{
Summarise PSA results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$summarise_PSA()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-draw_plots"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-draw_plots}{}}}
\subsection{Method \code{draw_plots()}}{
Draw plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$draw_plots(
  prior_sample_method = "LHS",
  print_pair_correlations = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prior_sample_method}}{Sampling method used to generate prior
samples.}

\item{\code{print_pair_correlations}}{Print pair-wise correlations to
console.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
Plot Goodness of fit function(s)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-draw_GOF_measure"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-draw_GOF_measure}{}}}
\subsection{Method \code{draw_GOF_measure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$draw_GOF_measure(
  .engine_ = "plotly",
  .gof_ = "LLK",
  .percent_sampled_ = 10,
  .n_samples_ = 10000,
  .points_ = FALSE,
  .true_points_ = FALSE,
  .greys_ = FALSE,
  .scale_ = NULL,
  .coloring_ = "fill",
  .legend_ = FALSE,
  .zoom_ = FALSE,
  .x_axis_lb_ = NULL,
  .x_axis_ub_ = NULL,
  .y_axis_lb_ = NULL,
  .y_axis_ub_ = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.engine_}}{Plotting engine, currently c("plotly", "ggplot2")}

\item{\code{.gof_}}{Goodness of fit (GOF) measure - fitness function. Either
"LLK" or "SEE" for the log-likelihood and sum-of-squared-errors GOF,
respectively.}

\item{\code{.percent_sampled_}}{.percent_sampled_ The fraction of LHS, RGS, or
FGS samples to select.}

\item{\code{.n_samples_}}{Number of Grid samples to plot log likelihood}

\item{\code{.points_}}{Boolean for whether to add scatter plot}

\item{\code{.true_points_}}{Boolean for whether to add "True set" to plots.}

\item{\code{.greys_}}{Boolean for whether to use a Grey scale in the plot.}

\item{\code{.scale_}}{The colour bar colour-scale. Available options are Greys,
YlGnBu, Greens, YlOrRd, Bluered, RdBu, Reds, Blues, Picnic, Rainbow,
Portland, Jet, Hot, Blackbody, Earth, Electric, Viridis, Cividis.}

\item{\code{.coloring_}}{Which contouring is required (default fill) and options
are "fill" | "heatmap" | "lines" | "none"}

\item{\code{.legend_}}{Boolean for whether to show a legend (default is FALSE).
This parameter also controls text labels in the opposite way.}

\item{\code{.zoom_}}{Boolean (default FALSE) for whether to limit the resulting
plot to the min() and max() of the two dimensional contour plots.}

\item{\code{.x_axis_lb_}}{Lower bound of the plot's x axis.}

\item{\code{.x_axis_ub_}}{Upper bound of the plot's x axis.}

\item{\code{.y_axis_lb_}}{Lower bound of the plot's y axis.}

\item{\code{.y_axis_ub_}}{Upper bound of the plot's y axis.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
Plot Targets

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-draw_targets_plots"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-draw_targets_plots}{}}}
\subsection{Method \code{draw_targets_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$draw_targets_plots(
  .engine_ = "ggplot2",
  .sim_targets_ = FALSE,
  .calibration_methods_ = c("random", "directed", "bayesian"),
  .legend_pos_ = "bottom",
  .PSA_samples_ = NULL,
  .PSA_unCalib_values_ = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.engine_}}{Plotting engine, currently c("plotly", "ggplot2")}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
Plot prior and posterior distributions

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-draw_distributions_plots"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-draw_distributions_plots}{}}}
\subsection{Method \code{draw_distributions_plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$draw_distributions_plots(
  .engine_ = "ggplot2",
  .legend_pos_ = "bottom",
  .log_scaled_ = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.engine_}}{String naming plotting package currently only supports
"ggplot2".}

\item{\code{.legend_pos_}}{String (default bottom) setting legend position.}

\item{\code{.log_scaled_}}{Boolean for whether to use log scale in the x axis.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calibR_R6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-calibR_R6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibR_R6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
