% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\alias{print.calibrate}
\alias{plot.calibrate}
\title{External probability calibration}
\usage{
calibrate(prob, y, method = c("pratt", "iso"), pos.class = NULL)

\method{print}{calibrate}(x, ...)

\method{plot}{calibrate}(x, ...)
}
\arguments{
\item{prob}{Vector of predicted probabilities.}

\item{y}{Vector of binary (i.e., 0/1) outcomes. If \code{y} is coded as
anything other than 0/1, then you must specify which of the two categories
represents the "positive" class (i.e., the class for which the probabilities
specified in \code{prob} correspond to) via the \code{pos.class} argument.}

\item{method}{Character string specifying which calibration method to use.
Current options include:
\describe{

  \item{\code{"pratt"}}{Pratt scaling.}

  \item{\code{"iso"}}{Isotonic (i.e., monotonic) calibration.}

}}

\item{pos.class}{Numeric/character string specifying which values in \code{y}
correspond to the "positive" class. Default is \code{NULL}. (Must be
specified) whenever \code{y} is not coded as 0/1.)}
}
\value{
A \code{"calibrate"} object, which is essentially a list with the
following components:
\describe{

  \item{\code{"probs"}}{A data frame containing two columns: \code{original}
  (the original probability estimates) and \code{calibrated} (the calibrated
  probability estimates).}

  \item{\code{"calibrater"}}{The calibration function (essentially a fitted
  model object) which can be used to calibrate new probabilities.}

  \item{\code{"bs"}}{The Brier score between \code{prob} and \code{y}.}

}
}
\description{
Validates predicted probabilities against a set of observed (binary)
outcomes.
}
