% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_colorimetric.R
\name{calibrate_colorimetric}
\alias{calibrate_colorimetric}
\title{Colorimetric assay calibration}
\usage{
calibrate_colorimetric(conc, abs, with.blank = TRUE, nrep = 1)
}
\arguments{
\item{conc}{The concentration of the standards.}

\item{abs}{The measured absorbance values of the standards. Must be in the same order as the concentrations. Can be a concatenation of replicates if each replicate sequence is in the same order as the enumerated concentrations.}

\item{with.blank}{Boolean. Whether or not a blank (buffer only) is included in the values.}

\item{nrep}{The number of replicates included.}
}
\description{
Generates a standard curve for colormetric assays to determine macromolecular concentration.
}
\examples{
calibrate_colorimetric( conc, abs, with.blank == TRUE, nrep = 1 )

}
