% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{calibrateModel_beyesian}
\alias{calibrateModel_beyesian}
\title{Calibrate models using Bayesian methods - employing local IMIS_()}
\usage{
calibrateModel_beyesian(
  .b_method = "SIR",
  .func,
  .args,
  .l_targets,
  .l_params,
  .samples,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Calibrate models using Bayesian methods - employing local IMIS_()
}
