% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_test_helpers.R
\name{calibTest_rosen}
\alias{calibTest_rosen}
\title{Rosenbrock function}
\usage{
calibTest_rosen(xx = c(1, 1), ...)
}
\arguments{
\item{xx}{function input, a vector of two or more. The length of (number
of values in) \code{xx} determines the dimensionality of the function.}

\item{...}{artefact to allow this function to work with the code that
was scripted for other goodness-of-fit functions}
}
\description{
Rosenbrock function
}
