% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfu.R
\name{cfu}
\alias{cfu}
\title{Calculate CFU/mL dilutions}
\usage{
cfu(od, od.df, count.cf = 8e+08, target.cpmL, target.vol)
}
\arguments{
\item{od}{The measured OD of the starter culture.}

\item{od.df}{The dilution factor of the OD measurement.}

\item{count.cf}{The correction factor that relates OD to CFU/mL. Defaults to 8E8.}

\item{target.cpmL}{The desired CFU/mL in the new culture.}

\item{target.vol}{The desired volume of the new culture.}
}
\description{
Calculate CFU/mL dilutions
}
\examples{
cfu()

}
