% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{check_icers_functions_inputs}
\alias{check_icers_functions_inputs}
\title{Check plot_cea_results_table Function, and Supporting Functions', Inputs}
\usage{
check_icers_functions_inputs(
  .df_outcomes_,
  .label_effects_,
  .label_costs_,
  .output_type_ = NULL,
  .output_format_ = NULL,
  .df_effects_ = NULL,
  .df_costs_ = NULL,
  .interventions_labels_ = NULL,
  .wtp_key_values_ = NULL,
  .show_nmb_ = NULL,
  .show_pce_ = NULL,
  .show_evpi_ = NULL,
  .show_dominance_ = NULL,
  .format_values_ = NULL,
  .currency_symbol_ = NULL,
  .show_diag_ = NULL
)
}
\arguments{
\item{.df_outcomes_}{Dataframe or tibble containing average costs and effects
data.}

\item{.label_effects_}{Character indicating the name of the column containing
the mean values of the effects. Default is \code{"QALYs"}.}

\item{.label_costs_}{Character indicating the name of the column containing
the mean values of the costs. Default is \code{"Costs"}.}

\item{.output_type_}{Character scalar that takes one of three options:
\code{"html"}, \code{"latex"} or \code{"dataframe"}. Default is \code{"dataframe"}. If \code{"html"}
or \code{"latex"} were passed to this argument, the function calls
\link{generate_gt_table} internally to generate the required output. The
\code{"html"} is suitable for shiny application; whereas, \code{"latex"}is suitable for
PDF document.}

\item{.output_format_}{Character scalar taking one of two options: \code{"wide"}
or \code{"long"} specifying whether the generated results table would be in a wide
or long format, respectively.}

\item{.df_effects_}{Dataframe or tibble containing the effects values from PSA.
The number of columns should equate to the number of interventions names
passed to \code{.interventions_labels_} while the number of rows should equate the
number of PSA simulations to be summarised. Also, the names of the columns in
\code{.df_effects_} should be identical to and appear in the same order to those of
the interventions' names in \code{.interventions_labels_}.}

\item{.df_costs_}{Dataframe or tibble containing the costs values from PSA.
The dimensions (rows and columns) and column names of \code{.df_costs_} should be
identical to those of \code{.df_effects_}.}

\item{.interventions_labels_}{Named Character vector containing the names of
all interventions. The names (of the elements (the labels) in the named
vector) should be identical to and in the same order of the column names in
the dataframes passed to the \code{.df_effects_} and \code{.df_costs_} arguments.}

\item{.wtp_key_values_}{Numeric vector specifying the willingness-to-pay (WTP)
values to be used in estimating the NMB and probability of the optimal choice
being cost-effective. This argument takes effect if \code{.show_nmb_} is \code{TRUE},
and accepts a maximum of two values. Default is \code{c(20000, 30000)}.}

\item{.show_nmb_}{Logical for whether to show the expected Net Monetary
Benefits (NMB) in the returned table. Requires \code{.wtp_key_values_} not to be
\code{NULL}.}

\item{.show_pce_}{Logical for whether to show the
probability-of-being-cost-effective in the returned table. Requires
\code{.wtp_key_values_} not to be \code{NULL}.}

\item{.show_evpi_}{Logical for whether to show individual Expected Value of
Perfect Information (EVPI) in the returned table. Requires \code{.wtp_key_values_}
not to be \code{NULL}.}

\item{.show_dominance_}{Logical for whether to show dominance information in
the returned table.}

\item{.format_values_}{Logical for whether to return a formatted table.
Default is \code{TRUE}.}

\item{.currency_symbol_}{Character scalar representing the Hex code of the
currency symbol to use in labeling relevant results. Default is \code{"\\u00A3"}
for Sterling (GBP). Use \code{"\\u20AC"}for Euros or \code{"\\u0024"} for US Dollars.}

\item{.show_diag_}{Logical for whether to keep calculations diagnostics
column.}
}
\value{
Returns nothing or prints error messages as execution is halted
when a check fails.
}
\description{
Check plot_cea_results_table Function, and Supporting Functions', Inputs
}
\examples{
\dontrun{
df_outcomes <- data.frame(
  'Intervention' = ShinyPSA::Hyperphosphatemia_PSA$treat \%>\% unname(),
  'QALYs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$e) \%>\% unname(),
  'Costs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$c) \%>\% unname()
)
check_icers_functions_inputs(
  df_outcomes = df_outcomes,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs",
  .df_effects_ = NULL,
  .df_costs_ = NULL,
  .interventions_labels_ = NULL
)
check_icers_functions_inputs(
  df_outcomes = NULL,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs",
  .df_effects_ = ShinyPSA::Hyperphosphatemia_PSA$e,
  .df_costs_ = ShinyPSA::Hyperphosphatemia_PSA$c,
  .interventions_labels_ = NULL
)
}
}
