% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem.mass.R
\name{chem.mass}
\alias{chem.mass}
\title{Calculates the mass any chemical to add to a solution to reach a target concentration}
\usage{
chem.mass(
  chemicals,
  mass,
  final.conc,
  final.vol,
  vol.unit = c("mL", "L"),
  conc.unit = c("M", "mM")
)
}
\arguments{
\item{chemicals}{The name of each chemical}

\item{mass}{The mass of the chemical}

\item{final.conc}{The final concentration of the solution in MOLAR units}

\item{final.vol}{The final volume of solution}

\item{vol.unit}{The desired units of output}
}
\description{
Calculates the mass any chemical to add to a solution to reach a target concentration
}
\examples{
chem.mass()

}
