% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EVPPI.R
\name{compute_EVPPI}
\alias{compute_EVPPI}
\title{Compute the Expected Value of Perfect Partial Information (EVPPI).}
\usage{
compute_EVPPI(
  .PSA_data,
  .df_effects_ = NULL,
  .df_costs_ = NULL,
  .df_params_ = NULL,
  .interventions_labels_ = NULL,
  .EVPI = NULL,
  .WTPs = NULL,
  .set = NULL,
  .set_names = NULL,
  .subset_ = FALSE,
  .MAICER_ = 30000,
  .currency_symbol_ = "£",
  .individual_evppi_ = TRUE,
  .discount_rate_ = 0.035,
  .evppi_population_ = NULL,
  .time_horion_ = NULL,
  .session = NULL
)
}
\arguments{
\item{.PSA_data}{A list of class shinyPSA that contains summary PSA
results.}

\item{.df_effects_}{A matrix containing the \code{effects} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.df_costs_}{A matrix containing the \code{costs} from PSA. Number of
\code{columns} is equal to the interventions while the number of
\code{rows} is equal to the number of PSA simulations to be summarised.}

\item{.df_params_}{A matrix containing parameters' configurations used in
the PSA.}

\item{.interventions_labels_}{A vector containing the names of all
interventions. If not provided or less names than needed is provided,
the function will generate generic names, for example
\code{intervention 1}.}

\item{.EVPI}{A data table containing EVPI data.}

\item{.WTPs}{A vector containing the Willingness-to-pay values over
which EVPI values were estimated.}

\item{.set}{A vector of parameters' names for conditional EVPPI.}

\item{.set_names}{A vector of parameter-names to be used for EVPPI.}

\item{.subset_}{Boolean for whether to estimate conditional EVPPI for a
subset of parameters.}

\item{.MAICER_}{The Maximum acceptable incremental cost-effectiveness
ratio (MAICER) to be considered in the summary table. Default value is
\code{30,000}.}

\item{.currency_symbol_}{A character, the units to associate with the
monitory values in the summary table. Default is sterling pounds
(GBP) \code{"\u00A3"}.}

\item{.individual_evppi_}{Logical (default \code{TRUE}) to return per
person EVPPI, otherwise population EVPPI will be reported.}

\item{.discount_rate_}{The discount rate used to discount future
affected populations.}

\item{.evppi_population_}{The size of the population that is annually
affected by the competing health technologies under evaluation.}

\item{.time_horion_}{The time expected to pass (in years) before the
interventions under consideration change (how long before the decision
under consideration become obsolete or requires updating).}

\item{.session}{Shiny app session.}
}
\value{
A list containing the EVPPI results table and caption
information.
}
\description{
Compute the Expected Value of Perfect Partial Information (EVPPI).
}
\examples{
\dontrun{
library(ShinyPSA)

# Summarise PSA results:
PSA_summary = summarise_PSA_(
  .df_effects_ = ShinyPSA::Brennan_1K_PSA$e,
  .df_costs_ = ShinyPSA::Brennan_1K_PSA$c,
  .df_params_ = ShinyPSA::Brennan_1K_PSA$p,
  .interventions = ShinyPSA::Brennan_1K_PSA$treats)

# Estimate EVPPI:
EVPPI_ind_res <- compute_EVPPI(
    .PSA_data = PSA_summary,
    .MAICER_ = 30000)

EVPPI_pop_res <- compute_EVPPI(
    .PSA_data = PSA_summary,
    .MAICER_ = 20000,
    .individual_evppi_ = FALSE,
    .evppi_population_ = 1000,
    .time_horion_ = 5)

# Estimate Conditional EVPPI:
cEVPPI_ind_res <- compute_EVPPI(
    .PSA_data = PSA_summary,
    .set = c(6, 14, 15, 16),
    .subset_ = TRUE,
    .MAICER_ = 30000)

cEVPPI_pop_res <- compute_EVPPI(
    .PSA_data = PSA_summary,
    .set_names = c("theta6", "theta14", "theta15", "theta16", "theta5"),
    .subset_ = TRUE,
    .MAICER_ = 20000,
    .individual_evppi_ = FALSE,
    .evppi_population_ = 1000,
    .time_horion_ = 5)
}
}
