% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{compute_relative_cea_values}
\alias{compute_relative_cea_values}
\title{Estimate Relative Values}
\usage{
compute_relative_cea_values(
  .df_combined_cea_results_,
  .interventions_names_,
  .simulated_truth_label_,
  .measures_ = "NMB@",
  .currency_symbol_ = "£"
)
}
\arguments{
\item{.df_combined_cea_results_}{Dataframe containing the outputs from the
\code{plot_cea_results_table} function. This dataframe is expected to contain the
cost-effectiveness results from one calibration method in addition to the
simulated truth. It is also expected that this dataframe includes at least
three columns, two of which are named \code{"Method"} and \code{" "}, whereas the other
columns are named after the interventions under analysis.}

\item{.interventions_names_}{Character vector specifying the names of the
interventions under analysis. It is expected that the names of the
interventions exist as the column names in \code{.df_combined_cea_results_}.}

\item{.simulated_truth_label_}{Character scalar specifying the label used
under the 'Method' column in \code{.df_combined_cea_results_} to indicate which
rows represent values from the simulated truth.}

\item{.measures_}{Character vector that take any combination of \code{"NMB@"},
\code{"PCE@"} or \code{"EVPI@"}.}

\item{.currency_symbol_}{Character scalar representing the Hex code of the
currency symbol to use in labeling relevant results. Default is \code{"\\u00A3"}
for Sterling (GBP). Use \code{"\\u20AC"}for Euros or \code{"\\u0024"} for US Dollars.}
}
\description{
This function replaces the absolute Net Monetary Benefit (NMB), Probability
of Cost-Effectiveness (PCE), and Expected Value of Perfect Information (EVPI)
resulting post calibration with the relative values compared to the simulated
truth.
}
\examples{
\dontrun{
compute_relative_cea_values(
 .df_combined_cea_results_ = df_PSA_combo_cea_tab_calibs,
 .interventions_names_ = .calibR_object_$model_interventions$v_interv_names,
 .simulated_truth_label_ = "Truth",
 .measures_ = "NMB@",
 .currency_symbol_ = "\u00A3"
 )
}
}
