% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution.calc.R
\name{dilution.calc}
\alias{dilution.calc}
\title{Calculates the volume of each component in a dilution}
\usage{
dilution.calc(
  samples,
  final.vol,
  final.conc,
  starting.conc,
  combine = c("yes", "no")
)
}
\arguments{
\item{samples}{The name of each sample being diluted}

\item{final.vol}{The targeted final volume}

\item{final.conc}{The targeted final concentration}

\item{starting.conc}{The starting concentration}

\item{combine}{Whether or not the final mixture should be a mixture of two analytes (e.g. for a binding reaction)}
}
\description{
Calculates the volume of each component in a dilution
}
\examples{
dilution.calc()

}
