% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Estimate Effective Sample Size from Bayesian calibration outputs}
\usage{
effective_sample_size(
  bayes_calib_output_list,
  .results = "Results",
  .post_prob_ = "Posterior_prob"
)
}
\arguments{
\item{bayes_calib_output_list}{List of outputs from the Bayesian
calibration}

\item{.results}{Name of calibration PSA draws table in the list}

\item{.post_prob_}{Name of column where posterior probability are found}
}
\description{
Estimate Effective Sample Size from Bayesian calibration outputs
}
\examples{
\dontrun{
}
}
