% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{generate_cea_results_tables}
\alias{generate_cea_results_tables}
\title{Generate Cost-Effectiveness Analysis Results Tables}
\usage{
generate_cea_results_tables(
  .ls_calibration_results_,
  .ls_PSA_results_,
  .model_interventions_,
  .model_outcomes_,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs",
  .wtp_key_values_ = c(20000, 30000),
  .highlight_optimal_choices_ = FALSE,
  .currency_symbol_ = "£",
  .output_type_ = "html",
  .full_output_format_ = "long",
  .add_simulated_truth_ = TRUE,
  .truth_PSA_output_list_path_ = "../5. Presentations/data/CRS_true_PSA.rds",
  .truth_PSA_output_list_ = NULL,
  .generate_partial_cea_table_ = TRUE,
  .partial_cea_table_groups_ = c("NMB", "PCE", "EVPI"),
  .generate_relative_values_ = TRUE,
  .relative_values_data_ = c("NMB")
)
}
\arguments{
\item{.ls_calibration_results_}{List containing the PSA of the calibration
results.}

\item{.ls_PSA_results_}{List containing the costs and effects PSA outputs.}

\item{.model_interventions_}{Character Vector specifying the names of the
interventions associated with the decision problem and outputs of the model.}

\item{.model_outcomes_}{Character vector specifying the labels of the
outcomes produced by the model, usually "costs" and "effects" or "QALYs".}

\item{.label_effects_}{Character indicating the name of the column containing
the mean values of the effects. Default is \code{"QALYs"}.}

\item{.label_costs_}{Character indicating the name of the column containing
the mean values of the costs. Default is \code{"Costs"}.}

\item{.wtp_key_values_}{Numeric vector specifying the willingness-to-pay
(WTP) values to be used in estimating the NMB and probability of the optimal
choice being cost-effective. Default is \code{c(20000, 30000)}.}

\item{.highlight_optimal_choices_}{Logical for whether to report NMB, PCE or
EVPI for the optimal choice. Default is \code{TRUE}. If \code{FALSE}, the values of
NMB, PCE or EVPI will be reported for all options.}

\item{.currency_symbol_}{Character scalar representing the Hex code of the
currency symbol to use in labeling relevant results. Default is \code{"\\u00A3"}
for Sterling (GBP). Use \code{"\\u20AC"}for Euros or \code{"\\u0024"} for US Dollars.}

\item{.output_type_}{Character scalar that takes one of three options:
\code{"html"}, \code{"latex"} or \code{"dataframe"}. Default is \code{"dataframe"}. If \code{"html"}
or \code{"latex"} were passed to this argument, the function calls
\link{generate_gt_table} internally to generate the required output. The
\code{"html"} is suitable for shiny application; whereas, \code{"latex"}is suitable for
PDF document.}

\item{.full_output_format_}{Character scalar taking one of two options:
\code{"wide"} or \code{"long"} specifying whether the generated results table would be
in a wide or long format, respectively. Default is \code{"long"}.}

\item{.add_simulated_truth_}{Logical scalar for whether to add results from
the Simulated Truth PSA data to the generated results tables. Default is
\code{TRUE}.}

\item{.truth_PSA_output_list_path_}{Character scalar specifying the path
where the list containing the Simulated Truth PSA data. The path is expected
to lead to a \code{.RDS} file representing the list containing the PSA data. This
list is expected to contain two dataframes named "e" or "effects" and "c" or
"costs", containing the effects and costs PSA data, respectively; and a
character vector containing the names of the interventions. This argument is
ignored if \code{.add_simulated_truth_} is set to \code{FALSE}.}

\item{.truth_PSA_output_list_}{List containing the Simulated Truth PSA data.
This list is expected to contain two dataframes named "e" or "effects" and
"c" or "costs", containing the effects and costs PSA data, respectively; and
a character vector containing the names of the interventions. This argument
is ignored if \code{.add_simulated_truth_} is set to \code{FALSE} or if
\code{.truth_PSA_output_list_path_} is set to \code{NULL}.}

\item{.generate_partial_cea_table_}{Logical for whether to generate CEA
tables with fewer results. Default is \code{TRUE}.}

\item{.partial_cea_table_groups_}{Character vector specifying what CEA
measures to report on the tables with fewer results. Default is
\code{c("NMB", "PCE", "EVPI")}, but it can take any or all of these values.}

\item{.generate_relative_values_}{Logical for whether the produced CEA
results table should be relative to the Simulated Truth or not. Default is
\code{TRUE}.}

\item{.relative_values_data_}{Character vector specifying what CEA measures
to report on the tables with fewer results. Default is \code{c("NMB")}, but it can
take any or all of\code{c("NMB", "PCE", "EVPI")}.}
}
\description{
Draw Cost-Effectiveness Analysis Results Tables from Probabilistic
Sensitivity Analysis (PSA) outputs. These PSA outputs are expected to be post
one or more calibration processes and possibly the simulated truth.
}
\examples{
\dontrun{
PSA_summary_tables1 <- generate_cea_results_tables(
   .ls_calibration_results_ = CR_CRS_2P1T$calibration_results,
   .ls_PSA_results_ = CR_CRS_2P1T$PSA_results,
   .model_interventions_ = CR_CRS_2P1T$model_interventions$v_interv_names,
   .model_outcomes_ = CR_CRS_2P1T$model_interventions$v_interv_outcomes,
   .label_effects_ = "QALYs",
   .label_costs_ = "Costs",
   .wtp_key_values_ = c(20000, 30000),
   .currency_symbol_ = "\u00A3",
   .output_type_ = "html",
   .full_output_format_ = "long",
   .add_simulated_truth_ = TRUE,
   .truth_PSA_output_list_path_ = "../5. Presentations/data/CRS_true_PSA.rds",
   .truth_PSA_output_list_ = NULL,
   .generate_partial_cea_table_ = TRUE,
   .partial_cea_table_groups_ = c("NMB", "PCE", "EVPI"),
   .generate_relative_values_ = TRUE,
   .relative_values_data_ = c("NMB")
)
PSA_summary_tables2 <- generate_cea_results_tables(
   .ls_calibration_results_ = CR_CRS_2P1T$calibration_results,
   .ls_PSA_results_ = CR_CRS_2P1T$PSA_results,
   .model_interventions_ = CR_CRS_2P1T$model_interventions$v_interv_names,
   .model_outcomes_ = CR_CRS_2P1T$model_interventions$v_interv_outcomes,
   .label_effects_ = "QALYs",
   .label_costs_ = "Costs",
   .wtp_key_values_ = c(20000, 30000),
   .currency_symbol_ = "\u00A3",
   .output_type_ = "html",
   .full_output_format_ = "long",
   .add_simulated_truth_ = TRUE,
   .truth_PSA_output_list_path_ = "../5. Presentations/data/CRS_true_PSA.rds",
   .truth_PSA_output_list_ = NULL,
   .generate_partial_cea_table_ = TRUE,
   .partial_cea_table_groups_ = c("NMB", "PCE", "EVPI"),
   .generate_relative_values_ = TRUE,
   .relative_values_data_ = c("NMB", "PCE")
)
PSA_summary_tables3 <- generate_cea_results_tables(
   .ls_calibration_results_ = CR_CRS_2P1T$calibration_results,
   .ls_PSA_results_ = CR_CRS_2P1T$PSA_results,
   .model_interventions_ = CR_CRS_2P1T$model_interventions$v_interv_names,
   .model_outcomes_ = CR_CRS_2P1T$model_interventions$v_interv_outcomes,
   .label_effects_ = "QALYs",
   .label_costs_ = "Costs",
   .wtp_key_values_ = c(20000, 30000),
   .currency_symbol_ = "\u00A3",
   .output_type_ = "html",
   .full_output_format_ = "long",
   .add_simulated_truth_ = TRUE,
   .truth_PSA_output_list_path_ = "../5. Presentations/data/CRS_true_PSA.rds",
   .truth_PSA_output_list_ = NULL,
   .generate_partial_cea_table_ = TRUE,
   .partial_cea_table_groups_ = c("NMB", "PCE", "EVPI"),
   .generate_relative_values_ = TRUE,
   .relative_values_data_ = c("NMB", "PCE", "EVPI")
)
}
}
