% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSA_helpers.R
\name{generate_PSA_samples}
\alias{generate_PSA_samples}
\title{Sample Probabilistic Sensitivity Analysis (PSA) values}
\usage{
generate_PSA_samples(.l_params, .PSA_samples, .transform_)
}
\arguments{
\item{.l_params}{A list that contains a vector of parameter names,
distributions and distributions' arguments.}

\item{.PSA_samples}{The number of PSA samples to sample or tabulate.}

\item{.transform_}{Logical, if TRUE the back transformation functions
in the .l_params list will be used to transform the parameters to their
original scale.}
}
\value{
A dataframe containing the PSA parameter configuration.
}
\description{
Sample Probabilistic Sensitivity Analysis (PSA) values
}
\examples{
\dontrun{
}
}
