% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_targets_helpers.R
\name{get_clean_method_name}
\alias{get_clean_method_name}
\title{Get cleaner calibration methods names}
\usage{
get_clean_method_name(.method_)
}
\arguments{
\item{.method_}{String naming the calibration method name to be cleaned.}
}
\description{
Get cleaner calibration methods names
}
\examples{
\dontrun{
}
}
