% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitness_helpers.R
\name{get_gof_contour_plot_title}
\alias{get_gof_contour_plot_title}
\title{Create contour plot title based on the calibration method and its outputs}
\usage{
get_gof_contour_plot_title(
  .gof_function_,
  .calib_method_,
  .colors_,
  .points_names
)
}
\arguments{
\item{.gof_function_}{String specifying the name of the goodness-of-fit (GOF)
measure used in generating the contour.}

\item{.calib_method_}{String specifying the name of the calibration method used
to generate the points displayed over the contour plot.}

\item{.colors_}{String vector naming the colours of the each of displayed
points.}

\item{.points_names}{String vector labelling the points displayed on the
contour plot.}
}
\description{
Create contour plot title based on the calibration method and its outputs
}
\examples{
\dontrun{
get_gof_contour_plot_title(
 .gof_function_ = "LLK",
 .calib_method_ = "IMIS",
 .colors_ = c(`Sampled sets` = "black",
             `Starting sets` = "black",
             `Prior samples` = "grey",
             `Identified sets` = "red",
             `Local extremas` = "darkorange",
             `Global extrema` = "red",
             `PSA draws` = "pink",
             `Posterior samples` = "red",
             `Posterior centres` = "yellow"),
 .points_names = c("Posterior centres",
                   "Posterior samples",
                   "Prior samples",
                   "True set")
 )
}
}
