% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_targets_helpers.R
\name{get_target_plot_title}
\alias{get_target_plot_title}
\title{Create plot title based on the calibration method and its outputs}
\usage{
get_target_plot_title(
  .scale_names_ = scale_names,
  .scale_colors_ = scale_colors,
  .target_ = .target_,
  .method_ = .calib_method_
)
}
\arguments{
\item{.scale_names_}{String vector specifying the levels or names of line
groups in the plot.}

\item{.scale_colors_}{String vector identifying the colours hex codes for
each of the groups in the plot.}

\item{.target_}{String naming the calibration target.}

\item{.method_}{String naming the calibration method that generate the
results in the plot.}
}
\description{
Create plot title based on the calibration method and its outputs
}
\examples{
\dontrun{
}
}
