% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibson.R
\name{gibson}
\alias{gibson}
\title{Calculate Gibson assembly reaction volumes}
\usage{
gibson(
  insert.conc,
  insert.len,
  vec.conc,
  vec.len,
  n.frags,
  molar.ratio = 3,
  vec.mass = 50,
  final.vol = 10
)
}
\arguments{
\item{insert.conc}{Vector of insert concentrations.}

\item{insert.len}{Vector of insert lengths in bp.}

\item{vec.conc}{Vector concentration. Must be a single value.}

\item{vec.len}{Vector length. Must be a single value.}

\item{molar.ratio}{Molar ratio of insert:vector}

\item{vec.mass}{Mass of vector}

\item{final.vol}{Final volume of the reaction before addition of 2x master mix.}
}
\description{
Calculate Gibson assembly reaction volumes
}
\examples{
gibson()

}
