% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HID_MarkovModel.R
\name{HID_markov}
\alias{HID_markov}
\title{Hypothetical Infectious Disease (HID) Markov Model}
\usage{
HID_markov(
  .v_params = NULL,
  project_future = FALSE,
  mu_e = 0.05,
  mu_l = 0.25,
  mu_t = 0.025,
  p = 0.1,
  r_l = 0.5,
  rho = 0.025,
  b = 0.2
)
}
\arguments{
\item{.v_params}{a named vector of model parameters in the following
order: "mu_e", "mu_l", "mu_t", "p", "r_l", "r_e", "rho", "b", "c".}

\item{project_future}{TRUE/FALSE, whether to project future outcomes for
policy comparison}

\item{pop_size}{Population size default is 1 million}

\item{mu_b}{Background mortality default is 0.015}
}
\value{

}
\description{
Hypothetical Infectious Disease (HID) Markov Model
}
