% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HID_MarkovModel.R
\name{HID_markov_2}
\alias{HID_markov_2}
\title{Hypothetical Infectious Disease (HID) Markov Model with transformation}
\usage{
HID_markov_2(
  .v_params_ = NULL,
  calibrate_ = TRUE,
  transform_ = TRUE,
  mu_e = log(0.04),
  mu_l = log(0.15),
  mu_t = log(0.016),
  p = log(0.12),
  r_l = log(0.41),
  rho = log(0.53),
  b = calibR::prob_to_logit(0.21)
)
}
\arguments{
\item{.v_params_}{a named vector of model parameters in the following
order: "mu_e", "mu_l", "mu_t", "p", "r_l", "r_e", "rho", "b", "c".}

\item{calibrate_}{Logical scalar, if \code{TRUE} (default), the model outputs
natural history data; otherwise, discounted outcomes (costs and QALYs)
are returned.}

\item{transform_}{Logical scalar, for whether to expect transformed
parameters and therefore back-transform them before they go into the model.}

\item{mu_e}{Cause-specific mortality rate with early-stage disease}

\item{mu_l}{Cause-specific mortality rate with late-stage disease}

\item{mu_t}{Cause-specific mortality rate on treatment}

\item{p}{Transition rate from early to late-stage disease}

\item{r_l}{Rate of uptake onto treatment (r_l = late-stage disease)}

\item{rho}{Effective contact rate}

\item{b}{Fraction of population in at-risk group}
}
\description{
Hypothetical Infectious Disease (HID) Markov Model with transformation
}
\examples{
\dontrun{
}
}
