% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{identify_all_e_dominance}
\alias{identify_all_e_dominance}
\title{Identify all extendedly (weakly) dominated interventions}
\usage{
identify_all_e_dominance(
  .df_outcomes_,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
\arguments{
\item{.df_outcomes_}{Dataframe or tibble containing average costs and effects
data.}

\item{.label_effects_}{Character indicating the name of the column containing
the mean values of the effects. Default is \code{"QALYs"}.}

\item{.label_costs_}{Character indicating the name of the column containing
the mean values of the costs. Default is \code{"Costs"}.}
}
\value{
A dataframe or table containing the information included in
\code{.df_outcomes_} in addition to information about which interventions were
extendedly dominated.
}
\description{
This function iteratively employs the \link{identify_e_dominance} and
\link{calculate_icers} functions to identify all waekly dominated
interventions.
}
\examples{
\dontrun{
df_outcomes <- data.frame(
  'Intervention' = ShinyPSA::Hyperphosphatemia_PSA$treat \%>\% unname(),
  'QALYs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$e) \%>\% unname(),
  'Costs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$c) \%>\% unname()
)
df_icer <- calculate_icers(
  .df_outcomes_ = df_outcomes,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
df_e_dominance <- identify_all_e_dominance(
  .df_outcomes_ = df_icer,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
df_icer2 <- calculate_icers(
  .df_outcomes_ = df_e_dominance,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
}
