% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{identify_dominance}
\alias{identify_dominance}
\title{Identify dominated interventions}
\usage{
identify_dominance(
  .df_outcomes_,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
\arguments{
\item{.df_outcomes_}{Dataframe or tibble containing average costs and effects
data.}

\item{.label_effects_}{Character indicating the name of the column containing
the mean values of the effects. Default is \code{"QALYs"}.}

\item{.label_costs_}{Character indicating the name of the column containing
the mean values of the costs. Default is \code{"Costs"}.}
}
\value{
A dataframe or table containing intervention names, mean costs and
effects corresponding to each intervention and information about which
interventions were strongly dominated.
}
\description{
This function identifies dominated interventions, ones that are more
effective but cheaper compared to the decision options ranked by effects.
Strong dominance rules out any of the competing interventions if there is
another intervention that is both more effective and less costly.
}
\examples{
\dontrun{
df_outcomes <- data.frame(
  'Intervention' = ShinyPSA::Hyperphosphatemia_PSA$treat \%>\% unname(),
  'QALYs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$e) \%>\% unname(),
  'Costs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$c) \%>\% unname()
)
df_dominance1 <- identify_dominance(
  .df_outcomes_ = df_outcomes,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
df_dominance2 <- identify_dominance(
  .df_outcomes_ = df_dominance1,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
}
