% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_summary_objects.R
\name{identify_e_dominance}
\alias{identify_e_dominance}
\title{Identify extendedly (weakly) dominated interventions}
\usage{
identify_e_dominance(
  .df_outcomes_,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
\arguments{
\item{.df_outcomes_}{Dataframe or tibble containing average costs and effects
data.}

\item{.label_effects_}{Character indicating the name of the column containing
the mean values of the effects. Default is \code{"QALYs"}.}

\item{.label_costs_}{Character indicating the name of the column containing
the mean values of the costs. Default is \code{"Costs"}.}
}
\value{
A dataframe or table containing the information included in
\code{.df_outcomes_} in addition to information about which interventions were
extendedly dominated.
}
\description{
This function employs the \link{calculate_icers} functions to identify
waekly dominated interventions, interventions with Incremental Cost
Effectiveness Ratio (ICER) higher than more expensive/effective
interventions.
Extended dominance (weak dominance) rules out any intervention that has an
incremental cost-effectiveness ratio that is greater than that of a more
effective intervention.
}
\examples{
\dontrun{
df_outcomes <- data.frame(
  'Intervention' = ShinyPSA::Hyperphosphatemia_PSA$treat \%>\% unname(),
  'QALYs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$e) \%>\% unname(),
  'Costs' = colMeans(ShinyPSA::Hyperphosphatemia_PSA$c) \%>\% unname()
)
df_icer <- calculate_icers(
  .df_outcomes_ = df_outcomes,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
df_e_dominance <- identify_e_dominance(
  .df_outcomes_ = df_icer,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
df_icer2 <- calculate_icers(
  .df_outcomes_ = df_e_dominance,
  .label_effects_ = "QALYs",
  .label_costs_ = "Costs"
)
}
}
