% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIS_.R
\name{IMIS_}
\alias{IMIS_}
\title{Incremental Mixture Importance Sampling (IMIS) - amended version}
\usage{
IMIS_(
  B = 1000,
  B.re = 3000,
  number_k = 100,
  D = 0,
  sample.prior = .sample.prior_,
  prior = .prior_,
  likelihood = .likelihood_
)
}
\arguments{
\item{prior}{}
}
\value{

}
\description{
Incremental Mixture Importance Sampling (IMIS) - amended version
}
