% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fonts.R
\name{import_plots_fonts}
\alias{import_plots_fonts}
\title{Download and install Google fonts.}
\usage{
import_plots_fonts(.name_ = "Source Sans Pro")
}
\arguments{
\item{.name_}{}
}
\description{
Download and install Google fonts.
}
