% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectious_disease_stm.R
\name{infectious_disease_stm}
\alias{infectious_disease_stm}
\title{Infectious Disease Markov Model with transformation}
\usage{
infectious_disease_stm(
  .v_params_ = NULL,
  calibrate_ = TRUE,
  transform_ = TRUE,
  mu_e = log(0.04),
  mu_l = log(0.15),
  mu_t = log(0.016),
  p = log(0.12),
  r_l = log(0.41),
  rho = log(0.53),
  b = calibR::prob_to_logit(0.21)
)
}
\arguments{
\item{.v_params_}{Named numeric vector of model parameters. If not named,
then the expected order is: \code{"mu_e"}, \code{"mu_l"}, \code{"mu_t"}, \code{"p"}, \code{"r_l"},
\code{"r_e"}, \code{"rho"}, \code{"b"}, \code{"c"}. Default is \code{NULL}.}

\item{calibrate_}{Logical scalar, if \code{TRUE} (default), the model outputs
natural history data; otherwise, discounted outcomes (costs and QALYs)
are returned.}

\item{transform_}{Logical scalar, for whether to expect transformed
parameters and therefore back-transform them before they go into the model.
Default is \code{TRUE}.}

\item{mu_e}{Numeric scalar representing the cause-specific mortality rate
with early-stage disease. Default model value is \code{0.04 [0.036, 0.044]} or
\code{log(0.04)} if \code{transform_} is \code{TRUE}.}

\item{mu_l}{Numeric scalar representing the cause-specific mortality rate
with late-stage disease. Default model value is \code{0.15 [0.13, 0.17]} or
\code{log(0.15)} if \code{transform_} is \code{TRUE}.}

\item{mu_t}{Numeric scalar representing the cause-specific mortality rate on
treatment. Default model value is \code{0.016 [0.013, 0.018]} or \code{log(0.016)} if
\code{transform_} is \code{TRUE}.}

\item{p}{Numeric scalar representing the transition rate from early to
late-stage disease. Default model value is \code{0.12 [0.09, 0.15]} or
\code{log(0.12)} if \code{transform_} is \code{TRUE}.}

\item{r_l}{Numeric scalar representing the rate of uptake onto treatment for
those at late-stage disease. Default model value is \code{0.41 [0.35, 0.48]} or
\code{log(0.41)} if \code{transform_} is \code{TRUE}.}

\item{rho}{Numeric scalar representing the effective contact rate. Default
model value is \code{0.53 [0.50, 0.55]} or \code{log(0.53)} if \code{transform_} is \code{TRUE}.}

\item{b}{Numeric scalar representing the fraction of population in at-risk
group. Default model value is \code{0.21 [0.18, 0.23]} or
\code{calibR::prob_to_logit(0.21)} if \code{transform_} is \code{TRUE}.}
}
\description{
Infectious Disease Markov Model with transformation
}
\examples{
\dontrun{
CE_results <- infectious_disease_stm()
}
}
