% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{log_prior}
\alias{log_prior}
\title{Calculate log prior}
\usage{
log_prior(.samples, .l_params)
}
\arguments{
\item{.samples}{A vector/dataset containing sampled/proposed values.}

\item{.l_params}{A list that contains a vector of parameter names,
distributions and distributions' arguments.}
}
\value{

}
\description{
Calculate log prior
}
