% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{logit_to_prob}
\alias{logit_to_prob}
\title{Convert logit to probability}
\usage{
logit_to_prob(.logit_)
}
\arguments{
\item{.logit_}{The logit to be transformed}
}
\value{
The probability corresponding to the passed logit
}
\description{
Convert logit to probability
}
