% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{lognormal_to_normal}
\alias{lognormal_to_normal}
\title{Convert Log-Normal Distribution Parameters to Normal Distribution Parameters}
\usage{
lognormal_to_normal(.mean_log_, .sd_log_)
}
\arguments{
\item{.mean_log_}{Numeric scalar representing the Mean of the Log-Normal
distribution.}

\item{.sd_log_}{Numeric scalar representing the Standard Deviation of the
Log-Normal distribution.}
}
\value{
A list containing the Mean and Standard Deviation of the Normal
distribution corresponding to those of the Log-Normal distribution
characterised by the inputs.
}
\description{
https://konvexity.com/relationship-between-normal-and-lognormal-distributions
https://rdrr.io/github/davidski/collector/src/R/fit_distributions.R
}
\examples{
\dontrun{
lognormal_to_normal(
 .mean_log_ = 1,
 .sd_log_ = 3
 )
}
}
