% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_to_radius.R
\name{mass_to_radius}
\alias{mass_to_radius}
\title{Convert molecular weight values to hydrodynamic radius}
\usage{
mass_to_radius(masses)
}
\arguments{
\item{mass}{The void volume of the column.}
}
\description{
Convert molecular weight values to hydrodynamic radius
}
\examples{
calibrate_sec( masses, void, cv, vols )

}
