% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HID_MarkovModel.R
\name{name_HID_params}
\alias{name_HID_params}
\title{Helper function to set the hypothetical infectious disease (HID) model}
\usage{
name_HID_params(.v_params)
}
\arguments{
\item{.v_params}{an unamed vector of model parameters in the following
order: "mu_e", "mu_l", "mu_t", "p", "r_l", "r_e", "rho", "b", "c".}
}
\value{

}
\description{
Helper function to set the hypothetical infectious disease (HID) model
}
