% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{normal_to_lognormal}
\alias{normal_to_lognormal}
\title{Convert Normal Distribution Parameters to Log-Normal Distribution Parameters}
\usage{
normal_to_lognormal(.mean_norm_, .sd_norm_)
}
\arguments{
\item{.mean_norm_}{Numeric scalar representing the Mean of the Normal
distribution.}

\item{.sd_norm_}{Numeric scalar representing the Standard Deviation of the
Normal distribution.}
}
\value{
A list containing the Mean and Standard Deviation of the Normal
distribution corresponding to those of the Log-Normal distribution
characterised by the inputs.
}
\description{
https://konvexity.com/relationship-between-normal-and-lognormal-distributions
https://rdrr.io/github/davidski/collector/src/R/fit_distributions.R
}
\examples{
\dontrun{
normal_to_lognormal(
 .mean_norm_ = 20,
 .sd_norm_ = 3
 )
}
}
