% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_ev.R
\name{normalize_ev}
\alias{normalize_ev}
\title{Normalize SEC elution volumes}
\usage{
normalize_ev(vols, void, cv)
}
\arguments{
\item{void}{The void volume of the column.}

\item{cv}{The column volume.}
}
\description{
Normalize SEC elution volumes to be between 0 and 1 using measured void volume and column volume
}
\examples{
calibrate_sec( masses, void, cv, vols )

}
