% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitness_helpers.R
\name{plot_fitness_contour}
\alias{plot_fitness_contour}
\title{Plot fitness function contour}
\usage{
plot_fitness_contour(
  .l_gof_values_,
  .l_params_,
  .gof_name_,
  .param_x,
  .param_y,
  .scale_,
  .greys_,
  .legend_,
  .coloring_,
  .true_points_,
  .legend_x_t_llk_ = "0.15",
  .legend_x_t_sse_ = "0.15",
  .legend_x_nt_llk_ = "0.25",
  .legend_x_nt_sse_ = "0.25",
  .x_axis_lb_,
  .x_axis_ub_,
  .y_axis_lb_,
  .y_axis_ub_
)
}
\arguments{
\item{.l_gof_values_}{List containing the goodness-of-fit (GOF) values
corresponding to specific parameter(s) configurations.}

\item{.l_params_}{List containing information about calibration parameters,
including parameters' names, distributions, and boundaries.}

\item{.gof_name_}{String specifying the GOF function name.}

\item{.param_x}{String identifying the parameter plotted in the x-axis.}

\item{.param_y}{String identifying the parameter plotted in the y-axis.}

\item{.scale_}{String specifying colour-scale applied to the "ploty" contour.
The options are "Blackbody", "Bluered", "Blues", "Cividis", "Earth",
"Electric", "Greens", "Greys", "Hot", "Jet", "Picnic", "Portland", "Rainbow",
"RdBu", "Reds", "Viridis" (default), "YlGnBu", and "YlOrRd".}

\item{.greys_}{Logical for whether to use the "Greys" colour-scale. The
.scale_ parameter overrides .greys_ if it was not NULL.}

\item{.legend_}{Logical for whether to show the "plotly" contour colour-bar
legend.}

\item{.coloring_}{String specifying where the colour-scale set by the .scale_
parameter is applied on the "plotly" contour. The options are "fill",
"heatmap", "none", and "lines". The "fill" option (default) paints the colour
scales over the layers of the contour; the "heatmap" option employs a heatmap
gradient colouring between each contour level; the "none" option paints no
colours on the contour layers and uses a single colour with the contour
lines; and the "lines" option applies the colour-scale on the contour lines.}

\item{.true_points_}{Logical for whether to show "True values" on the plot.}

\item{.legend_x_t_llk_}{Numeric (double) value specifying the x-axis
relative location when "True values" are plotted and in "LLK" GOF plots.}

\item{.legend_x_t_sse_}{Numeric (double) value specifying the x-axis
relative location when "True values" are plotted and in "SSE" GOF plots.}

\item{.legend_x_nt_llk_}{Numeric (double) value specifying the x-axis
relative location when no "True values" are plotted and in "LLK" GOF plots.}

\item{.legend_x_nt_sse_}{Numeric (double) value specifying the x-axis
relative location when no "True values" are plotted and in "SSE" GOF plots.}

\item{.x_axis_lb_}{Numeric (double) value specifying the lower bound of x-axis.}

\item{.x_axis_ub_}{Numeric (double) value specifying the upper bound of x-axis.}

\item{.y_axis_lb_}{Numeric (double) value specifying the lower bound of y-axis.}

\item{.y_axis_ub_}{Numeric (double) value specifying the upper bound of y-axis.}
}
\description{
Plot fitness function contour
}
\examples{
\dontrun{
}
}
