% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors_posterior_helpers.R
\name{plot_pri_post_distributions}
\alias{plot_pri_post_distributions}
\title{Plot prior-posterior distributions}
\usage{
plot_pri_post_distributions(
  .engine_ = "ggplot2",
  .l_PSA_samples_ = self$PSA_samples,
  .l_params_ = self$calibration_parameters,
  .l_calibration_results_ = self$calibration_results,
  .t_prior_samples_ = self$prior_samples[["LHS"]],
  .transform_ = self$transform_parameters,
  .bins_ = 100,
  .legend_pos_ = "bottom",
  .log_scaled_ = FALSE
)
}
\arguments{
\item{.engine_}{String specifying the plotting engine. Currently supports
"ggplot2". Also, some of the code is written for the "triliscope" engine.}

\item{.l_PSA_samples_}{List containing PSA samples - from which the function
uses Bayesian calibration results.}

\item{.l_params_}{List containing information about calibration parameters,
including parameters' names, distributions, and boundaries.}

\item{.l_calibration_results_}{List containing the results from the
calibration methods.}

\item{.t_prior_samples_}{Dataset or tibble containing prior samples.}

\item{.transform_}{Logical for whether the model is set to handle parameters
on a transformed scale.}

\item{.bins_}{Numeric specifying the number of bins in the histograms.}

\item{.legend_pos_}{String defining the location of the legend position
default (bottom).}

\item{.log_scaled_}{Logical for whether to present the x-axis using the
log scale.}
}
\description{
Plot prior-posterior distributions
}
\examples{
\dontrun{
}
}
