% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{prob_to_logit}
\alias{prob_to_logit}
\title{Convert probability to logit}
\usage{
prob_to_logit(.prob_)
}
\arguments{
\item{.prob_}{The probability to be transformed}
}
\value{
The logit transformation of the passed probability
}
\description{
Convert probability to logit
}
