% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot.param.R
\name{prot.param}
\alias{prot.param}
\title{Calculates standard physical prameters of a protein given its amino acid sequence.}
\usage{
prot.param(aa.sequence, temp)
}
\arguments{
\item{aa.sequence}{The amino acid sequence of the protein of interest in single letter code. Enter as \code{c()}.}

\item{temp}{The temperature of the analysis. Relevant to calculation of various hydrodynamic parameters like vbar.}
}
\description{
Calculates standard physical prameters of a protein given its amino acid sequence.
}
\examples{
prot.param()

}
