% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_param.R
\name{prot_param}
\alias{prot_param}
\title{Calculates standard physical prameters of a protein given its amino acid sequence.}
\usage{
prot_param(aa.sequence, temp = NULL, wavelength = NULL)
}
\arguments{
\item{aa.sequence}{The amino acid sequence of the protein of interest in single letter code. Enter as \code{c()}.}

\item{temp}{The temperature of the analysis.}

\item{wavelength}{The wavelength being used by the refractometer.}
}
\description{
Calculates standard physical prameters of a protein given its amino acid sequence.
}
\examples{
prot_param()

}
