% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSA_helpers.R
\name{PSA_calib_values}
\alias{PSA_calib_values}
\title{Merge calibration outputs and sample PSA values}
\usage{
PSA_calib_values(
  .l_optim_lists = l_optim_lists,
  .search_method = "Directed",
  .PSA_runs = 1000,
  .l_params = NULL
)
}
\arguments{
\item{.l_params}{}
}
\value{

}
\description{
Merge calibration outputs and sample PSA values
}
\examples{
l_optim_lists <- list(GA_optimise_lLLK, GA_optimise_wSSE,
                      GB_optimise_lLLK, GB_optimise_wSSE,
                      NM_optimise_lLLK, NM_optimise_wSSE,
                      SA_optimise_lLLK, SA_optimise_wSSE)

calibrated_values <- extract_calib_values()

}
