% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_helpers.R
\name{run_demo_App}
\alias{run_demo_App}
\title{Run the example shiny app.}
\usage{
run_demo_App(example_app = "one")
}
\arguments{
\item{example_app}{The example shiny app to run.}
}
\value{
Runs the example shiny app
}
\description{
Run the example shiny app.
}
\examples{
\dontrun{
run_API_demo_App()
}
}
