% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSA_helpers.R
\name{run_PSA}
\alias{run_PSA}
\title{Run Probabilistic Sensitivity Analysis (PSA) for a Decision Analytic
Models (DAM) using both calibrated and un-calibrated parameters.}
\usage{
run_PSA(.func_, .args_, .PSA_calib_values_, .PSA_unCalib_values_)
}
\arguments{
\item{.func_}{The Decision Analytic Models (DAM) that was calibrated and
for which Probabilistic Sensitivity Analysis (PSA) will be conducted.}

\item{.args_}{Extra arguments passed to .func_, the DAM.}

\item{.PSA_calib_values_}{PSA values sampled for the calibration
parameter(s).}

\item{.PSA_unCalib_values_}{PSA values sampled for un-calibrated
(known) parameters(s).}
}
\value{

}
\description{
Run Probabilistic Sensitivity Analysis (PSA) for a Decision Analytic
Models (DAM) using both calibrated and un-calibrated parameters.
}
\examples{
\dontrun{
}
}
