% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_prior.R
\name{sample_prior_LHS}
\alias{sample_prior_LHS}
\title{Use Latin Hypercube Sampling (LHS) to sample from prior distribution}
\usage{
sample_prior_LHS(.n_samples = 1, .l_params = .l_params_, ...)
}
\arguments{
\item{.n_samples}{An integer specifying the number of samples to be
generated.}

\item{.l_params}{A list that contains a vector of parameter names,
distributions and distributions' arguments.}

\item{...}{additional arguments, for example: .seed_no to set a seed
number.}
}
\value{
A table with each parameter LHS samples in a separate column
}
\description{
Use Latin Hypercube Sampling (LHS) to sample from prior distribution
}
\examples{
\dontrun{
v_params_names <- c("p_Mets", "p_DieMets")
v_params_dists <- c("unif", "unif")
args <- list(list(min = 0.04, max = 0.16),
             list(min = 0.04, max = 0.12))
l_params <- list('v_params_names' = v_params_names,
                 'v_params_dists' = v_params_dists,
                 'args' = args)

sample_prior_LHS(.l_params = l_params,
                 .n_samples = 10)
}
}
