% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_helpers.R
\name{summ_optim}
\alias{summ_optim}
\title{Summarise and/or create uncertainty estimates from the hessian matrix}
\usage{
summ_optim(
  .params_name = v_params_names,
  .gof = NULL,
  .gof_name,
  .gof_value,
  .s_method,
  .par,
  .func = NULL,
  .args = NULL,
  .hessian = NULL,
  .maximiser = TRUE,
  ...
)
}
\arguments{
\item{.params_name}{Character vector containing the names of the
parameters that were passed to the goodness-of-fit algorithm or will be
passed to .func.}

\item{.gof}{goodness-of-fit function used or to be used in the
optimisation.}

\item{.gof_name}{Character naming goodness-of-fit method that produced
samples}

\item{.gof_value}{Numeric goodness-of-fit value for the corresponding
parameters.}

\item{.s_method}{Character naming search algorithm that produced
.GoF_value.}

\item{.par}{Parameter values to be used to generate \code{95% confidence
interval} and/or passed to .func to generate the .hessian matrix.}

\item{.func}{A function passed to and to be optimised by .gof.}

\item{.args}{A list of arguments passed to .func.}

\item{.hessian}{The hessian matrix.}

\item{.maximiser}{Logical for whether algorithm that created (or .func
which will create) the hessian matrix maximised the goodness-of-fit
function. Default is \code{TRUE}.}

\item{...}{Extra arguments to be passed to .gof.}
}
\value{
A tibble with the best identified parameters, their 95\%
confidence intervals and corresponding goodness-of-fit values.
}
\description{
Summarise and/or create uncertainty estimates from the hessian matrix
}
