% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCalibCurve.R
\name{survCalibCurve}
\alias{survCalibCurve}
\title{Calibration Analysis of Survival Predictions}
\usage{
survCalibCurve(data, time, event, p, timepoint, showCStat = TRUE,
  p.out = NULL, na.rm = FALSE, bs.knots = c(0.01, 0.025, 0.05, 0.1, 0.2,
  0.3), bs.degree = 3)
}
\arguments{
\item{data}{A dataset}

\item{time}{unquoted name of the time variable}

\item{event}{unquoted name of the event variable (0/1)}

\item{p}{unquoted name of the variable containing predicted probabilities (at
t=timepoint)}

\item{timepoint}{Value of 'time' where probabilities are extracted (from the
Kaplan-Meier curves).}

\item{showCStat}{Logical. If TRUE, a concordance statistic (from
\code{rms::somers2()}) with confidence limits (2 std. errors) will be displayed}

\item{p.out}{grid of predicted probability values at which calibration curve
is estimated. If NULL, a sequence of 10 values evenly spaced along the range
of inputted predicted probabilities is generated.}

\item{na.rm}{Logical. Remove rows with missing values?}

\item{bs.degree}{degree parameter passed to \code{splines::bs()}}

\item{knots}{knots parameter passed to \code{splines::bs()}}
}
\description{
Calibration Analysis of Survival Predictions
}
\details{
The predicted probabilities are transform via the logit prior to
spline curve fitting, in order to stabilize the resulting calibration
curve estimates.  Therefore, if specifying the \code{knots=} parameter to
\code{bs.params}, the values must be expressed on the logit scale. (It's
usually easier/sufficient to just supply \code{df} instead of \code{knots}.)

Spline fits in Cox proportional hazards model are prone to minor convergence
issues, which impact the validity of Wald-based standard errors, tests, and
confidence intervals. Hence this may produce a warning of the form "Loglik
converged before variable X; beta may be infinite". See
https://stat.ethz.ch/pipermail/r-help/2008-September/174201.html for details.

The user is advised to choose the knots for the splines very carefully,
especially in the presence of highly skewed distributions of model-predicted
probabilities. The default values are just a suggestion for a typical
scenario where predictions are highly skewed to the right.
}
