% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv_vis.R
\name{uv_vis}
\alias{uv_vis}
\title{Determine macromolecule concentration from UV/Vis measurements.}
\usage{
uv_vis(
  abs,
  ext = NULL,
  df = 1,
  plen = 1,
  type = c("protein", "dsDNA", "ssDNA", "ssRNA")
)
}
\arguments{
\item{abs}{The measured absorbance value at the wavelength of interest.}

\item{ext}{The extinction coefficient of the molecule. The units of the extinction coefficient determine the units of concentration.}

\item{df}{The dilution factor of the measurement.}

\item{plen}{The path length of the measurement. Defaults to 1 cm.}

\item{type}{The type of molecule being analyzed.}
}
\description{
Determine macromolecule concentration from UV/Vis measurements.
}
\examples{
uv_vis()

}
