% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_helpers.R
\name{calibrateModel_beyesian}
\alias{calibrateModel_beyesian}
\title{Calibrate models using Bayesian methods - employing local IMIS_()}
\usage{
calibrateModel_beyesian(
  .b_method = "SIR",
  .func,
  .args,
  .l_targets,
  .l_params,
  .samples,
  .n_resample = 1000,
  .IMIS_sample = 1000,
  .IMIS_iterations = 30,
  .IMIS_optimiser = TRUE,
  .MCMC_burnIn = 10000,
  .MCMC_samples = 50000,
  .MCMC_thin = 5,
  .MCMC_rerun = TRUE,
  .transform = FALSE,
  .diag_ = FALSE
)
}
\arguments{
\item{.b_method}{Character defining the Bayesian method to use in the
calibration process. Currently supported methods are \code{SIR} and \code{IMIS}.}

\item{.func}{A function defining the decision analytic model to be
calibrated.}

\item{.args}{A list of arguments passed to the model function.}

\item{.l_targets}{A list containing a vector of targets' names, a vector
of targets' weights, a vector of targets' distributions, and a table for
each target that contains the values (column name 'value') and standard
errors (column name 'sd') of the corresponding target.}

\item{.l_params}{A list that contains a vector of parameter names,
distributions and distributions' arguments.}

\item{.samples}{A table or vector of sampled parameter values}

\item{.n_resample}{the desired number of draws from the posterior}

\item{.IMIS_sample}{the incremental sample size at each IMIS iteration}

\item{.IMIS_iterations}{the maximum number of iterations in IMIS}

\item{.IMIS_optimiser}{logical for whether to employ optimisation algorithms
within the IMIS method}

\item{.MCMC_burnIn}{the number of samples before starting to retain samples}

\item{.MCMC_samples}{the total number of samples the MCMC algorithm should
generate including the burn-in sample including the .MCMC_burnIn. This value
should not be equal to or less than .MCMC_burnIn.}

\item{.MCMC_thin}{the value used to thin the resulting chain}

\item{.MCMC_rerun}{use the proposal distribution covariance matrix from the
first run to re-run the MCMC chain.}

\item{.transform}{Logical for whether to back-transform parameters to
their original scale.}

\item{.diag_}{Logical for whether to print diagnostics}
}
\description{
Calibrate models using Bayesian methods - employing local IMIS_()
}
\examples{
\dontrun{
}
}
