% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateR_R6.R
\name{calibrateR_R6}
\alias{calibrateR_R6}
\title{R6 Class representing a calibrateR_R6 calibration machine.}
\format{
An \link[R6:R6Class]{R6::R6Class} object.
}
\description{
An instance of this class is expected to calibrate a user-defined model
and produce summary plots/tables.
}
\examples{
\dontrun{
}


## ------------------------------------------------
## Method `calibrateR_R6$sampleR`
## ------------------------------------------------

\dontrun{
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{calibration_model}}{a Decision-Analytic model under calibration}

\item{\code{prior_samples}}{a Decision-Analytic model under calibration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sampleR}{\code{calibrateR_R6$sampleR()}}
\item \href{#method-clone}{\code{calibrateR_R6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sampleR"></a>}}
\if{latex}{\out{\hypertarget{method-sampleR}{}}}
\subsection{Method \code{sampleR()}}{
Initialisation method (triggered when a new object is created).
Summary plots and table(s) are created alongside the construction
of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrateR_R6$sampleR(.n_samples = 1, .l_params = .l_params_, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.n_samples}}{}

\item{\code{.l_params}}{}

\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{calibrateR_R6} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{calibrateR_R6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
