% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibson.R
\name{gibson}
\alias{gibson}
\title{Calculates the volumes of vector and insert of a given series of Gibson assembly reactions}
\usage{
gibson(
  insert.list,
  vector.list,
  insert.conc,
  insert.len,
  vec.conc,
  vec.len,
  molar.ratio,
  final.vol,
  vec.mass = NULL,
  single.reaction = c("yes", "no")
)
}
\arguments{
\item{insert.list}{List of all inserts being put into plasmids}

\item{vector.list}{List of all vectors being cloned into. Must enter one vector for each insert, even if there are duplicates.}

\item{insert.conc}{The concentration of insert in ng/µL}

\item{insert.len}{The insert size in bp}

\item{vec.conc}{The concentration of the vector in ng/µL}

\item{vec.len}{The linearized vector size in bp}

\item{molar.ratio}{The molar ratio of insert to vector}

\item{final.vol}{The final volume of the reaction BEFORE addition of 2x reaction buffer.}

\item{vec.mass}{The target mass of vector. If none given, assumed to be 50 ng.}

\item{single.reaction}{\code{"yes"} or \code{"no"}. Tells the function whether or not to treat all entries as a single reaction or multiple reactions.}
}
\description{
Calculates the volumes of vector and insert of a given series of Gibson assembly reactions
}
\examples{
gibson()

}
