% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HID_MarkovModel.R
\name{HID_markov}
\alias{HID_markov}
\title{Hypothetical Infectious Disease (HID) Markov Model}
\usage{
HID_markov(
  .v_params_ = NULL,
  calibrate_ = TRUE,
  mu_e = 0.05300305,
  mu_l = 0.09492525,
  mu_t = 0.06949119,
  p = 0.35494867,
  r_l = 0.70669255,
  rho = 0.62436759,
  b = 0.27229941
)
}
\arguments{
\item{.v_params_}{a named vector of model parameters in the following
order: "mu_e", "mu_l", "mu_t", "p", "r_l", "r_e", "rho", "b", "c".}

\item{calibrate_}{If \code{TRUE} (default), the model outputs natural
history data; otherwise, discounted outcomes \code{(costs and QALYs)}
are returned.}

\item{mu_e}{Cause-specific mortality rate with early-stage disease}

\item{mu_l}{Cause-specific mortality rate with late-stage disease}

\item{mu_t}{Cause-specific mortality rate on treatment}

\item{p}{Transition rate from early to late-stage disease}

\item{r_l}{Rate of uptake onto treatment (r_l = late-stage disease)}

\item{rho}{Effective contact rate}

\item{b}{Fraction of population in at-risk group}
}
\value{

}
\description{
Hypothetical Infectious Disease (HID) Markov Model
}
\examples{
\dontrun{
}
}
