% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HID_MarkovModel.R
\name{HID_markov_2}
\alias{HID_markov_2}
\title{Hypothetical Infectious Disease (HID) Markov Model with transformations}
\usage{
HID_markov_2(
  .v_params_ = NULL,
  calibrate_ = TRUE,
  transform_ = TRUE,
  mu_e = log(0.05),
  mu_l = log(0.25),
  mu_t = log(0.025),
  p = log(0.1),
  r_l = log(0.5),
  rho = log(0.025),
  b = prob_to_logit(0.2)
)
}
\arguments{
\item{.v_params_}{a named vector of model parameters in the following
order: "mu_e", "mu_l", "mu_t", "p", "r_l", "r_e", "rho", "b", "c".}

\item{calibrate_}{If \code{TRUE} (default), the model outputs natural
history data; otherwise, discounted outcomes \code{(costs and QALYs)}
are returned.}

\item{transform_}{Expect transformed parameters and therefore
back-transform them before they go into the model}

\item{mu_e}{Cause-specific mortality rate with early-stage disease}

\item{mu_l}{Cause-specific mortality rate with late-stage disease}

\item{mu_t}{Cause-specific mortality rate on treatment}

\item{p}{Transition rate from early to late-stage disease}

\item{r_l}{Rate of uptake onto treatment (r_l = late-stage disease)}

\item{rho}{Effective contact rate}

\item{b}{Fraction of population in at-risk group}
}
\value{

}
\description{
Hypothetical Infectious Disease (HID) Markov Model with transformations
}
