% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_targets_helpers.R
\name{plot_targets}
\alias{plot_targets}
\title{Plot observed and simulated targets}
\usage{
plot_targets(
  .engine_ = "ggplot2",
  .l_targets_ = self$calibration_targets,
  .simulated_targets_ = self$simulated_targets,
  .sim_targets_ = FALSE,
  .legend_pos_ = "bottom"
)
}
\arguments{
\item{.engine_}{String naming the plotting engine, currently "ggplot2".}

\item{.l_targets_}{List containing calibration (observed) targets
information.}

\item{.simulated_targets_}{List containing simulated targets.}

\item{.sim_targets_}{Logical (default FALSE) for whether to generate
plots for the simulated targets.}

\item{.legend_pos_}{String defining the location of the legend position
default (bottom).}
}
\description{
Plot observed and simulated targets
}
\examples{
\dontrun{
targets_plots <- plot_targets(
  .l_targets_ = CR_CRS_2P2T$calibration_targets,
  .simulated_targets_ = CR_CRS_2P2T$simulated_targets,
  .sim_targets_ = TRUE)
}
}
