% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\name{sec}
\alias{sec}
\title{Gel filtration column calibration and analyte analysis}
\usage{
sec(void, cv, stds, masses, unk, sample.ids)
}
\arguments{
\item{void}{The void volume of the column, measured with something like blue dextran.}

\item{cv}{The column volume of the column, measured with something like 0.2\% (v/v) acetone.}

\item{stds}{The elution volumes of the standards, entered within \code{c()}.}

\item{masses}{The known masses of each of the standards.}

\item{unk}{The elution volume of any unknown analyte whose hydrodynamic parameters are of interest. Required for any \code{type="pred"} usage.}
}
\description{
Gel filtration column calibration and analyte analysis
}
\examples{
sec(masses, void, cv, stds, unk)

}
