% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv_vis.R
\name{uv_vis}
\alias{uv_vis}
\title{Calculates the concentration of protein, dsDNA, ssDNA, or ssRNA from absorbance values at 280 or 260 nm}
\usage{
uv_vis(
  type = c("protein", "dsDNA", "ssDNA", "ssRNA"),
  A260,
  A280,
  dil.fac,
  path.length = NULL,
  extinction.coeff = NULL,
  mol.weight = NULL
)
}
\arguments{
\item{type}{The type of molecule being analyzed \code{protein},\code{dsDNA},\code{ssDNA},or \code{ssRNA}}

\item{A260}{The absorbance value at 260 nm}

\item{A280}{The absorbance value at 280 nm}

\item{dil.fac}{The dilution factor of the measured sample}

\item{path.length}{The path length of the instrument in cm. Usually 1 cm.}

\item{extinction.coeff}{The extinction coefficient of a given protein in molar units.}

\item{mol.weight}{The molecular weight of a given protein. Necessary to return concentration in mg/mL.}
}
\description{
Calculates the concentration of protein, dsDNA, ssDNA, or ssRNA from absorbance values at 280 or 260 nm
}
\examples{
uv_vis()

}
